Fit a list of parsnip models

This allows for the simple fitting of multiple models. It requires three components: a formula, a parsnip model definition, and a data set to analyze.

fit_parsnip_models(.formula, .test, .opts = NULL, .data)

Arguments

.formula
List or vector of formulas

.test
A model definition from parsnip

.opts
Options to pass to the test, if needed

.data
Data set to be used

Value

Returns a list of model fits

See also